refactor: 반복되는 ResponseMetaData build 관련 로직을 static Generator로 분리
