+++ src/main/java/configure/ContentType.java
@@ -4,4 +4,7 @@ public class ContentType {
     public static final String HTML_CONTENT_TYPE = "text/html";
     public static final String CSS_CONTENT_TYPE = "text/css";
     public static final String JAVASCRIPT_CONTENT_TYPE = "text/javascript";
+
+    private ContentType() {
+    }
 }
