+++ src/main/java/controller/ResourceController.java
@@ -2,8 +2,8 @@ package controller;
 
 import webserver.request.HttpRequest;
 import webserver.response.HttpResponse;
-import webserver.response.HttpStatus;
 import webserver.response.ResponseMetaData;
+import webserver.response.ResponseMetaDataGenerator;
 
 import java.io.IOException;
 
@@ -11,10 +11,7 @@ public class ResourceController extends AbstractController {
 
     @Override
     public void service(final HttpRequest request, final HttpResponse response) throws IOException {
-        ResponseMetaData responseMetaData = ResponseMetaData.Builder
-                .builder(request, HttpStatus.OK)
-                .contentType(request.findContentType())
-                .build();
+        ResponseMetaData responseMetaData = ResponseMetaDataGenerator.buildDefaultOkMetaData(request);
         response.setResponseMetaData(responseMetaData);
 
         doGet(request, response);
