+++ src/main/java/controller/page/IndexPageController.java
@@ -1,20 +1,18 @@
-package controller;
+package controller.page;
 
+import controller.AbstractController;
 import webserver.request.HttpRequest;
 import webserver.response.HttpResponse;
-import webserver.response.HttpStatus;
 import webserver.response.ResponseMetaData;
+import webserver.response.ResponseMetaDataGenerator;
 
 import java.io.IOException;
 
-public class IndexController extends AbstractController {
+public class IndexPageController extends AbstractController {
 
     @Override
     public void service(final HttpRequest request, final HttpResponse response) throws IOException {
-        ResponseMetaData responseMetaData = ResponseMetaData.Builder
-                .builder(request, HttpStatus.OK)
-                .contentType(request.findContentType())
-                .build();
+        ResponseMetaData responseMetaData = ResponseMetaDataGenerator.buildDefaultOkMetaData(request);
         response.setResponseMetaData(responseMetaData);
 
         doGet(request, response);
